// models/content-types/conference.ts
import { type IContentItem, type Elements } from '@kontent-ai/delivery-sdk';
import { type Focus } from '../taxonomies/focus';

/**
 * Generated by '@kontent-ai/model-generator@5.8.0'
 *
 * Conference
 * Id: 62952827-a865-45c5-b884-b6fc1eb009d8
 * Codename: conference
 */
export type Conference = IContentItem<{
  /**
   * Address (text)
   * Required: true
   * Id: 0b9efbff-cf09-4a87-b0a4-2d15462e5701
   * Codename: address
   */
  address: Elements.TextElement;

  /**
   * Description (rich_text)
   * Required: false
   * Id: 4e39eb4a-68bb-4444-86b5-adcaa379f3b0
   * Codename: description
   */
  description: Elements.RichTextElement;

  /**
   * End date (date_time)
   * Required: false
   * Id: 0f815e31-4560-4b27-be78-3a9863ed6e66
   * Codename: end_date
   */
  end_date: Elements.DateTimeElement;

  /**
   * Focus (taxonomy)
   * Required: true
   * Id: 4d455ffc-b044-416c-afc8-abbff3c3f4db
   * Codename: focus
   */
  focus: Elements.TaxonomyElement<Focus>;

  /**
   * Name (text)
   * Required: true
   * Id: c2e79854-644b-45bc-9743-6a7e4c4ffa13
   * Codename: name
   */
  name: Elements.TextElement;

  /**
   * Start date (date_time)
   * Required: false
   * Id: 548f6ea8-6ab6-444c-be74-b220d13b4b16
   * Codename: start_date
   */
  start_date: Elements.DateTimeElement;

  /**
   * URL slug (url_slug)
   * Required: true
   * Id: c6dbf547-1adf-45c1-821c-6d57fbbd765f
   * Codename: url_slug
   */
  url_slug: Elements.UrlSlugElement;

  /**
   * Venue photo (asset)
   * Required: true
   * Id: 1f9fc8ac-bd14-4f56-a8be-34fe1f39fcf6
   * Codename: venue_photo
   */
  venue_photo: Elements.AssetsElement;
}>;